blog/requirements.txt
Django==4.2.7
pylint == 3.0.2
DjangoRestFramework == 3.14.0
drf-yasg==1.21.7
django-filter==23.5